Move Info classes to info package.
