+++ features-extraction/src/main/java/org/jetbrains/research/groups/ml_methods/extraction/Main.java
@@ -1,4 +1,4 @@
-package org.jetbrains.research.groups.ml_methods;
+package org.jetbrains.research.groups.ml_methods.extraction;
 
 public class Main {
     public static void main(String[] args) {
