+++ features-extraction/src/main/java/org/jetbrains/research/groups/ml_methods/extraction/info/InfoCollector.java
@@ -1,4 +1,4 @@
-package org.jetbrains.research.groups.ml_methods;
+package org.jetbrains.research.groups.ml_methods.extraction.info;
 
 import com.intellij.analysis.AnalysisScope;
 import com.intellij.psi.*;
